@@ -1,43 +1,96 @@
-import React from 'react';
-import { HashRouter, Route, Switch } from 'react-router-dom';
+import { useEffect, useState, useReducer } from "react";
+import { HashRouter, Route, Switch } from "react-router-dom";
+import { Helmet } from "react-helmet";
 
-import './App.css';
+import "./App.css";
 
-import HotelView from './views/HotelView';
-import MapView from './views/MapView';
-import AboutView from './views/AboutView';
-import ContactView from './views/ContactView';
-import PersonView from './views/PersonView';
-import DataExportView from './views/DataExportView';
-import DataImportView from './views/DataImportView';
+import HotelView from "./views/HotelView";
+import MapView from "./views/MapView";
+import ContentPageView from "./views/ContentPageView";
+import PersonView from "./views/PersonView";
+import ErrorView from "./views/ErrorView";
+import ErrorBoundary from "./components/ErrorBoundary";
 
-import { MapContext, HotelContext } from './context';
-import reducer, { initialState } from './reducer';
+import { MapContext, HotelContext } from "./context";
+import reducer, { initialState } from "./reducer";
+import { useTranslation } from "react-i18next";
+import { config } from "./config";
 
-import hotels from './data/nerhotel.json';
+import loadHotelDataFromCsv from "./utils/load-hotel-data-from-csv";
 
-function App () {
-  const [state, dispatch] = React.useReducer(reducer, initialState);
-  const mapData = { ...state, dispatch};
+function App() {
+  const [state, dispatch] = useReducer(reducer, initialState);
+  const mapData = { ...state, dispatch };
+  const { i18n, t } = useTranslation();
+  const [hotels, setHotels] = useState([]);
+
+  useEffect(() => {
+    const queryString = window.location.href.split("?")[1];
+    const urlParamsObj = new URLSearchParams(queryString);
+
+    let preferredLang;
+    if (urlParamsObj.has(config.locales.paramName)) {
+      preferredLang = urlParamsObj.get(config.locales.paramName);
+    } else {
+      preferredLang = localStorage.getItem(config.locales.paramName);
+    }
+
+    if (
+      preferredLang &&
+      config.locales.available.includes(preferredLang) &&
+      preferredLang !== i18n.options.lng
+    ) {
+      i18n.changeLanguage(preferredLang);
+      localStorage.setItem(config.locales.paramName, preferredLang);
+    }
+  }, [i18n]);
+
+  useEffect(() => {
+    let isSubscribed = true;
+
+    loadHotelDataFromCsv()
+      .then((data) => {
+        if (isSubscribed) {
+          setHotels(data);
+        }
+      })
+      .catch((e) => {
+        console.error(e);
+      })
+      .finally(() => {
+        dispatch({ type: "SetDataLoaded" });
+      });
+
+    return () => {
+      isSubscribed = false;
+    };
+  }, []);
 
   return (
-    <div className="App">
-      <HotelContext.Provider value={{hotels}}>
-        <MapContext.Provider value={mapData}>
-          <HashRouter>
-            <Switch>
-              <Route path="/" exact component={MapView}/>
-              <Route path="/hotel/:id" exact component={HotelView}/>
-              <Route path="/about" exact component={AboutView}/>
-              <Route path="/contact" exact component={ContactView}/>
-              <Route path="/person/:name" exact component={PersonView}/>
-              <Route path="/data-export" exact component={DataExportView}/>
-              <Route path="/data-import" exact component={DataImportView}/>
-            </Switch>
-          </HashRouter>
-        </MapContext.Provider>
-      </HotelContext.Provider>
-    </div>
+    <>
+      <ErrorBoundary>
+        <Helmet>
+          <title>
+            {t("general:tagline")} - {t("general:siteName")}
+          </title>
+        </Helmet>
+        <HotelContext.Provider value={{ hotels }}>
+          <MapContext.Provider value={mapData}>
+            <HashRouter>
+              <Switch>
+                <Route path="/" exact component={MapView} />
+                <Route path="/hotel/:id" exact component={HotelView} />
+                <Route path="/about" exact component={ContentPageView} />
+                <Route path="/contact" exact component={ContentPageView} />
+                <Route path="/person/:name" exact component={PersonView} />
+                <Route path="/data-export" exact component={ContentPageView} />
+                <Route path="*" component={ErrorView} />
+              </Switch>
+            </HashRouter>
+          </MapContext.Provider>
+        </HotelContext.Provider>
+      </ErrorBoundary>
+    </>
   );
 }
 
