@@ -1,4 +1,4 @@
-import React from 'react';
+import { useContext, useCallback } from 'react';
 
 import Search from './Search';
 import styles from '../css/header.module.css';
@@ -12,15 +12,30 @@ import restaurant from '../assets/nh-restaurant.svg';
 import golf from '../assets/nh-golf.svg';
 import { MapContext } from '../context';
 
+import { config } from '../config';
+import LangSwitch from './LangSwitch';
+
+import { useTranslation } from 'react-i18next';
+import { useLocation } from 'react-router-dom';
+
+import { SmartLink } from './SmartLink';
+
 const Header = (props) => {
-  const { dispatch } = React.useContext(MapContext);
+  const { dispatch } = useContext(MapContext);
+  const { i18n } = useTranslation();
+  const location = useLocation();
 
-  const goHome = () => {
-    dispatch({ type: 'TogglePopup', showPopup: false });
-    props.history.push('/');
-  };
+  const languageChangeHandler = (e) => {
+    const lang = e.target.value;
+    i18n.changeLanguage(lang);
+    localStorage.setItem(config.locales.paramName, lang);
+    props.history.push({
+      pathname: location.pathname,
+      search: `?${config.locales.paramName}=${lang}`,
+    });
+  }
 
-  const onMenuCallback = React.useCallback(() => {
+  const onMenuCallback = useCallback(() => {
     dispatch({ type: 'ToggleMenu', showMenu: true })
   }, [dispatch]);
 
@@ -32,9 +47,13 @@ const Header = (props) => {
         <Icon img={listIcon} size="large"/>
       </div>
       <div className={styles.headerWrapper}>
-        <div onClick={() => goHome()} className={styles.logo}>
-          <img src={logo} alt=""/>
-        </div>
+        <SmartLink
+          to="/"
+          onClick={() => { dispatch({ type: 'TogglePopup', showPopup: false }); }}
+          className={styles.logo}
+        >
+          <img src={logo} alt="" />
+        </SmartLink>
         <div className={styles.headerInner}>
           <div className={styles.icons}>
             <img src={hotel} alt=""/>
@@ -44,6 +63,13 @@ const Header = (props) => {
           </div>
           {props.withSearch && <Search/>}
         </div>
+      </div>
+      <div className={styles.langSwitchContainer}>
+        <LangSwitch
+          availableLocales={config.locales.available}
+          onLanguageChange={languageChangeHandler}
+          currentLocale={i18n.resolvedLanguage}
+        />
       </div>
     </header>
   );
