@@ -1,16 +1,14 @@
-import React from 'react';
 import styles from '../css/powered-by-vercel.module.css';
+import { SmartLink } from './SmartLink';
 
 function PoweredByVercel(props) {
     return (
-        <a
-            href={props.link}
-            target="_blank"
-            rel="noopener noreferrer"
+        <SmartLink
+            to={props.link}
             className={styles.poweredByVercel}
         >
             Powered by ▲ Vercel
-        </a>
+        </SmartLink>
     );
 }
 
