@@ -1,55 +1,61 @@
 export const initialState = {
   list: [],
+  isDataLoaded: false,
   map: null,
   showList: false,
   showPopup: false,
   showMenu: false,
   center: [47.498045, 19.0385183],
   selectedPoint: null,
-  locationRequired: false
 };
 
 export default function mapReducer(state, action) {
+  if (process.env.NODE_ENV === "development") {
+    console.log(`==== ${action.type} ====`);
+    console.log("state: ", state);
+    console.log("action: ", action);
+    console.log(`========`);
+  }
   switch (action.type) {
-    case 'SetMap':
+    case "SetMap":
       return {
         ...state,
-        map: action.map
+        map: action.map,
       };
-    case 'SetList':
+    case "SetList":
       return {
         ...state,
-        list: action.list
+        list: action.list,
       };
-    case 'SetCenter':
+    case "SetCenter":
       return {
         ...state,
-        center: action.center
+        center: action.center,
       };
-    case 'SetSelectedPoint':
+    case "SetSelectedPoint":
       return {
         ...state,
-        selectedPoint: action.point
+        selectedPoint: action.point,
       };
-    case 'ToggleList':
+    case "ToggleList":
       return {
         ...state,
-        showList: action.showList
+        showList: action.showList,
       };
-    case 'TogglePopup':
+    case "TogglePopup":
       return {
         ...state,
-        showPopup: action.showPopup
+        showPopup: action.showPopup,
       };
-    case 'ToggleMenu':
+    case "ToggleMenu":
       return {
         ...state,
-        showMenu: action.showMenu
+        showMenu: action.showMenu,
       };
-    case 'SetLocator':
+    case "SetDataLoaded":
       return {
         ...state,
-        locationRequired: true
+        isDataLoaded: true,
       };
     default:
       return state;
