@@ -1,5 +1,5 @@
-const OWNER = 'cégtulajdonos';
-const MANAGER = 'cégvezető';
+const OWNER = "person:owner";
+const MANAGER = "person:manager";
 
 const reduceByName = (map, person) => {
   if (!map.has(person.name)) {
@@ -18,9 +18,16 @@ const reduceByName = (map, person) => {
  * @returns {{name: string, data: {name: string, link: string, type: string}}[]}
  */
 export function getOligarchData(oligarchs, ceos) {
-  const oligarchsWithType = oligarchs.map(person => ({...person, type: OWNER}));
-  const ceosWithType = ceos.map(person => ({...person, type: MANAGER}));
+  const oligarchsWithType = oligarchs.map((person) => ({
+    ...person,
+    type: OWNER,
+  }));
+  const ceosWithType = ceos.map((person) => ({ ...person, type: MANAGER }));
   const allOligarchsWithType = [...oligarchsWithType, ...ceosWithType];
   const oligarchMap = allOligarchsWithType.reduce(reduceByName, new Map());
-  return Array.from(oligarchMap.entries()).map(([name, data]) => ({name, data}));
+
+  return Array.from(oligarchMap.entries()).map(([name, data]) => ({
+    name,
+    data,
+  }));
 }
